requirements.txt
pyudev==0.22.0
evdev==1.3.0
paho-mqtt==1.6.1
tomli==2.0.1
pyzmq==24.0.1
chevron==0.14.0
